tweaks
git-svn-id: https://libsunflower.svn.sourceforge.net/svnroot/libsunflower/trunk/sunflower@19 86aca762-15d7-0310-9726-9bd8562223d8

+++ sfl/api/HAL.hpp
@@ -72,7 +72,7 @@ namespace sfl {
     virtual int speed_set(double qdl, double qdr) = 0;
     
     /** \return 0 on success. */
-    virtual int scan_get(int channel, double * rho, size_t rho_len,
+    virtual int scan_get(int channel, double * rho, int rho_len,
 			 struct ::timespec * t0, struct ::timespec * t1) = 0;
   };
